requirements.txt
Flask==2.3.3
mysql-connector-python==8.1.0
Flask-Login==0.6.3
python-dotenv==1.0.0
gunicorn==21.2.0
cryptography==41.0.4
reportlab==4.0.4
WTForms==3.0.1
